% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate_and_label_peaks.R
\name{annotate_and_label_peaks}
\alias{annotate_and_label_peaks}
\title{Annotate and label a set of genomic peaks}
\usage{
annotate_and_label_peaks(peaks, set_name, txdb, tss_region = c(-3000, 3000))
}
\arguments{
\item{peaks}{A GRanges object representing peaks to annotate.}

\item{set_name}{Character string labeling the source of the peaks.}

\item{txdb}{A TxDb object (e.g. from GenomicFeatures::makeTxDbFromGFF).}

\item{tss_region}{A numeric vector of length 2 specifying the TSS region.
Default: c(-3000, 3000)}
}
\value{
A data.frame of annotation statistics with a new column "Set"
}
\description{
Uses ChIPseeker::annotatePeak to annotate peaks and adds a column
indicating the peak set.
}
