% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculateLog2FcOverGRanges.R
\name{calculateLog2FcOverGRanges}
\alias{calculateLog2FcOverGRanges}
\title{Calculate log2 fold change of BAM signals over GRanges windows}
\usage{
calculateLog2FcOverGRanges(
  tx_bam_files,
  in_bam_files,
  gr,
  chromosomes_to_include = paste0("chr", 1:22),
  mapq = 15,
  n_cores = 8
)
}
\arguments{
\item{tx_bam_files}{Character vector of paths to treatment BAM files.}

\item{in_bam_files}{Character vector of paths to input BAM files.}

\item{gr}{A \code{GRanges} object representing regions over which to calculate CPM and log2FC.}

\item{chromosomes_to_include}{Character vector of chromosome names to include (default: paste0("chr", 1:22)).}

\item{mapq}{Integer. Minimum mapping quality to include (default: 15).}

\item{n_cores}{Number of cores for parallel processing (default: 8).}
}
\value{
A GRanges object with a \code{Log2FC} metadata column.
}
\description{
This function calculates the log2 fold change (log2FC) of coverage per million (CPM)
between treatment and input BAM files over a set of GRanges, typically corresponding
to windows such as those used in Repli-seq.
}
