% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_functions.R
\name{count_peaks_in_bins}
\alias{count_peaks_in_bins}
\title{Count Peaks Overlapping Genomic Bins}
\usage{
count_peaks_in_bins(peaks_gRanges, genomicRange_bins)
}
\arguments{
\item{peaks_gRanges}{A `GRanges` object containing peak locations.}

\item{genomicRange_bins}{A `GRanges` object representing genomic bins.}
}
\value{
A numeric vector of peak counts per bin.
}
\description{
This function counts the number of peaks that overlap with each genomic bin.
}
\examples{
bins <- create_bins_across_genomic_range("chr1", 100000, 200000, 5000)
peaks <- GRanges("chr1", IRanges(start = c(100100, 150000), width = 200))
count_peaks_in_bins(peaks, bins)
}
\seealso{
Other noble_peak_density_heatmap: 
\code{\link{convert_vector_to_dataframe_for_plot}()},
\code{\link{create_bins_across_genomic_range}()},
\code{\link{generate_peak_density_heatmap}()},
\code{\link{make_heatmap_of_peak_counts_in_bins}()}
}
\concept{noble_peak_density_heatmap}
