% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_bins_across_genomic_range.R
\name{create_bins_across_genomic_range}
\alias{create_bins_across_genomic_range}
\title{Create Genomic Bins Across a Range}
\usage{
create_bins_across_genomic_range(
  range_chromosome,
  range_start,
  range_end,
  binSize
)
}
\arguments{
\item{range_chromosome}{Character. The chromosome name (e.g., "chr1").}

\item{range_start}{Numeric. The start position of the genomic range.}

\item{range_end}{Numeric. The end position of the genomic range.}

\item{binSize}{Numeric. The width of each bin.}
}
\value{
A \code{GRanges} object containing genomic bins.
}
\description{
This function generates evenly spaced bins of a given size across a specified genomic range.
}
\examples{
create_bins_across_genomic_range("chr1", 100000, 200000, 5000)
}
\seealso{
Other noble_peak_density_heatmap: 
\code{\link{convert_vector_to_dataframe_for_plot}()},
\code{\link{count_peaks_in_bins}()},
\code{\link{generate_peak_density_heatmap}()},
\code{\link{make_heatmap_of_peak_counts_in_bins}()}
}
\concept{noble_peak_density_heatmap}
