% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_functions.R
\name{createAndWriteUnifiedPeakSet}
\alias{createAndWriteUnifiedPeakSet}
\title{Create a unified peak set from multiple peak files and write it to a file}
\usage{
createAndWriteUnifiedPeakSet(pkFiles, smtFiles, outputPath)
}
\arguments{
\item{pkFiles}{A list of paths to the peak files to be combined.}

\item{outputPath}{Path to the output file where the unified peak set will be saved.}
}
\value{
NULL (The function writes the result to a file and doesn't return a value.)
}
\description{
This function takes a list of peak files and creates a unified peak set
by combining the second set of peaks with non-overlapping peaks from the first set.
The resulting peak set is then written to a specified output file in BED format.
}
\examples{
pkFiles <- c("path/to/peakfile1.bed", "path/to/peakfile2.bed")
outputPath <- "path/to/output/unified_peaks"
createAndWriteUnifiedPeakSet(pkFiles, outputPath)

}
