% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/define_unified_reference_peaks.R
\name{define_unified_reference_peaks}
\alias{define_unified_reference_peaks}
\title{Define a Unified Reference Peak Set}
\usage{
define_unified_reference_peaks(peaks_gRanges_list)
}
\arguments{
\item{peaks_gRanges_list}{A list of \code{GRanges} objects, each representing a set of peaks.}
}
\value{
A \code{GRanges} object representing the unified peak set.
}
\description{
This function merges multiple peak sets into a single unified peak set,
reducing overlapping peaks into non-overlapping regions.
}
\examples{
peaks_list <- list(
  GRanges(seqnames="chr1", IRanges(start=c(100, 300), width=50)),
  GRanges(seqnames="chr1", IRanges(start=c(120, 350), width=50))
)
unified_peaks <- define_unified_reference_peaks(peaks_list)
print(unified_peaks)
}
\seealso{
\code{\link{make_peak_overlap_df}}, \code{\link{make_euler_plot_showing_peak_overlaps}}
}
