% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetReproduciblePeaks.R
\name{GetReproduciblePeaks}
\alias{GetReproduciblePeaks}
\title{Identify reproducible peaks across multiple replicates}
\usage{
GetReproduciblePeaks(queryPeaks, subjectPeaksList, overlapCountThreshold = 5)
}
\arguments{
\item{queryPeaks}{A \code{GRanges} object representing the set of query peaks (e.g., reproducible peaks).}

\item{subjectPeaksList}{A list of \code{GRanges} objects, typically representing peak calls from multiple replicates.}

\item{overlapCountThreshold}{Integer. Minimum number of replicates in which a peak must be present to be considered reproducible. Default is 5.}
}
\value{
A \code{GRanges} object containing only those peaks from \code{queryPeaks} that overlap at least \code{overlapCountThreshold} replicates.
}
\description{
This function identifies peaks from a query set that overlap with a minimum number of replicates.
It returns only those peaks from the query set that are present in at least \code{overlapCountThreshold}
of the GRanges objects in \code{subjectPeaksList}.
}
\examples{
# Example usage:
# reproduciblePeaks <- GetReproduciblePeaks(queryPeaks, listOfReplicatePeaks, overlapCountThreshold = 4)

}
