% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HCT_Repliseq_gr.R
\docType{data}
\name{HCT_Repliseq_gr}
\alias{HCT_Repliseq_gr}
\title{GenomicRanges with HCT116 Repliseq Matrix}
\format{
A GenomicRanges object.
}
\source{
GEO: GSE137764
}
\usage{
data(HCT_Repliseq_gr)
}
\description{
A GenomicRanges with matrix of values from high resolution repliseq.
}
\details{
All peaks are called with MACS2 and used to derive reproducible peaks in the analysis.
}
\keyword{datasets}
