% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hg38_blacklist_v2_granges.R
\docType{data}
\name{hg38_blacklist_v2_granges}
\alias{hg38_blacklist_v2_granges}
\title{GenomicRanges of hg38 blacklist from ENCODE (https://github.com/Boyle-Lab/Blacklist/blob/master/lists/hg38-blacklist.v2.bed.gz)}
\format{
A GenomicRanges object.
}
\usage{
data(hg38_blacklist_v2_granges)
}
\description{
GenomicRanges of hg38 blacklist from ENCODE (https://github.com/Boyle-Lab/Blacklist/blob/master/lists/hg38-blacklist.v2.bed.gz)
}
\keyword{datasets}
