% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R, R/treslin_peak_granges.R
\docType{data}
\name{hg38_ideogram_granges}
\alias{hg38_ideogram_granges}
\title{Human genome ideogram annotations for hg38}
\format{
A GRanges object with ideogram bands.

A GenomicRanges object.
}
\source{
UCSC Genome Browser
}
\usage{
hg38_ideogram_granges

data(hg38_ideogram_granges)
}
\description{
Ideogram band data for chromosomes in hg38, used in genome track visualizations.
}
\keyword{datasets}
