% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importBED.R
\name{importBED}
\alias{importBED}
\title{Import a BED file as a GenomicRanges object}
\usage{
importBED(bedPath, chromosomesToImport = chromosomes)
}
\arguments{
\item{bedPath}{Character string. Path to the BED file.}

\item{chromosomesToImport}{Character vector. A list of chromosomes to keep (default is \code{chromosomes}, assumed to be defined in the environment).}
}
\value{
A \code{GRanges} object containing only the specified chromosomes.
}
\description{
This function reads a BED file and converts it to a \code{GRanges} object,
retaining only the specified chromosomes.
}
\examples{
importBED("path/to/peaks.bed", chromosomesToImport = paste0("chr", 1:22))

}
