% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/make_euler_plot_of_overlaps_with_reference_peaks.R
\name{make_euler_plot_of_overlaps_with_reference_peaks}
\alias{make_euler_plot_of_overlaps_with_reference_peaks}
\title{Generate an Euler Plot of Peak Overlaps}
\usage{
make_euler_plot_of_overlaps_with_reference_peaks(peak_overlap_df, plot_colors)
}
\arguments{
\item{peak_overlap_df}{A data frame where each row represents a peak in the reference set,
and each column corresponds to a query peak set, with `TRUE`/`FALSE` values indicating overlaps.}

\item{plot_colors}{A vector of colors to use for the Euler plot.}
}
\value{
A `ggplot` object representing the Euler plot.
}
\description{
This function creates an Euler plot based on a data frame of peak overlaps.
}
\examples{
overlap_df <- data.frame(Sample1 = c(TRUE, FALSE, TRUE), Sample2 = c(FALSE, TRUE, TRUE))
colors <- c("#1b9e77", "#d95f02")
euler_plot <- make_euler_plot_of_overlaps_with_reference_peaks(overlap_df, colors)
print(euler_plot)
}
\seealso{
\code{\link{make_peak_overlap_df}}, \code{\link{make_euler_plot_showing_peak_overlaps}}
}
