% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_peak_overlap_df.R
\name{make_peak_overlap_df}
\alias{make_peak_overlap_df}
\title{Create a Data Frame Indicating Peak Overlaps}
\usage{
make_peak_overlap_df(subjectPeaks, queryPeaksList)
}
\arguments{
\item{subjectPeaks}{A \code{GRanges} object representing the reference peak set.}

\item{queryPeaksList}{A named list of \code{GRanges} objects, each representing a query peak set.}
}
\value{
A data frame where each row corresponds to a peak in \code{subjectPeaks},
and each column corresponds to a query peak set. Values are \code{TRUE} if the
subject peak overlaps the query peak set and \code{FALSE} otherwise.
}
\description{
This function takes a set of subject peaks and a list of query peak sets,
then determines whether each query peak set overlaps with the subject peaks.
}
\examples{
subjectPeaks <- GRanges(seqnames="chr1", IRanges(start=c(100,200), width=50))
queryPeaksList <- list(
  Sample1 = GRanges(seqnames="chr1", IRanges(start=c(120, 250), width=50)),
  Sample2 = GRanges(seqnames="chr1", IRanges(start=c(180, 300), width=50))
)
overlap_df <- make_peak_overlap_df(subjectPeaks, queryPeaksList)
print(overlap_df)
}
\seealso{
\code{\link{define_unified_reference_peaks}}, \code{\link{make_euler_plot_of_overlaps_with_reference_peaks}}
}
