% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/make_stacked_barchart_of_genomic_features_overlapping_peaks.R
\name{make_stacked_barchart_of_genomic_features_overlapping_peaks}
\alias{make_stacked_barchart_of_genomic_features_overlapping_peaks}
\title{Create a Stacked Bar Chart of Genomic Features Overlapping Peaks}
\usage{
make_stacked_barchart_of_genomic_features_overlapping_peaks(
  df_of_genomic_features_overlapping_peaks,
  feature_colors = c(`Promoter (<=1kb)` = "#9b2226", `Promoter (1-2kb)` = "#ae2012",
    `Promoter (2-3kb)` = "#bb3e03", `5' UTR` = "#ca6702", `3' UTR` = "#ee9b00",
    `1st Exon` = "#e9d8a6", `Other Exon` = "#F7F1DE", `1st Intron` = "#94d2bd",
    `Other Intron` = "#0a9396", `Downstream (<=300)` = "#005f73", `Distal Intergenic` =
    "#001219")
)
}
\arguments{
\item{df_of_genomic_features_overlapping_peaks}{A data frame containing the frequency of genomic features overlapping with peaks.
Expected columns:
\itemize{
\item \code{Feature}: Genomic feature (e.g., "Promoter", "Exon", "Intron").
\item \code{Frequency}: The number of peaks overlapping each feature.
\item \code{Set}: The label for the peak set.
}}

\item{feature_colors}{A named vector of colors for each genomic feature.
The default color scheme follows a standardized palette for promoters, UTRs, exons, introns, and intergenic regions.}
}
\value{
A \code{ggplot} object representing a stacked bar chart of genomic feature frequencies.
}
\description{
This function generates a stacked bar chart representing the frequency of genomic
features overlapping with peaks.
}
\examples{
df <- data.frame(
  Feature = c("Promoter (<=1kb)", "Exon", "Intron", "Distal Intergenic"),
  Frequency = c(100, 200, 300, 150),
  Set = c("Peaks Set 1", "Peaks Set 1", "Peaks Set 1", "Peaks Set 1")
)
make_stacked_barchart_of_genomic_features_overlapping_peaks(df)
}
\seealso{
\code{\link{make_df_of_genomic_features_overlapping_peaks}}
}
