% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeDomainHeatPlot.R
\name{makeDomainHeatPlot}
\alias{makeDomainHeatPlot}
\title{Create a Complex Heatmap of Domain Log2 Fold Change}
\usage{
makeDomainHeatPlot(
  domainsBed =
    "../GenerateTimingWindowsFromHCTRepliseq/TreslinDepletedLateRepDomains.bed",
  windowSize = 1e+05,
  stepSize = 5000,
  txBamFile1 =
    "../01_Asynchronous_HCT116/results/mergedDownSampledBams/TICRR_WT_HCT116_anti-GFP_250.bam",
  inBamFile1 =
    "../01_Asynchronous_HCT116/results/mergedDownSampledBams/WT_HCT116_anti-GFP_250.bam",
  txBamFile2 =
    "../01_Asynchronous_HCT116/results/mergedDownSampledBams/TICRR_WT_HCT116_anti-GFP_1000.bam",
  inBamFile2 =
    "../01_Asynchronous_HCT116/results/mergedDownSampledBams/WT_HCT116_anti-GFP_1000.bam",
  span = 3e+06
)
}
\arguments{
\item{domainsBed}{Path to the BED file containing genomic domains.}

\item{windowSize}{The size of the sliding window used for calculating Lg2FC.}

\item{stepSize}{The step size for sliding the window.}

\item{txBamFile1}{Path to the BAM file for treatment sample 1.}

\item{inBamFile1}{Path to the BAM file for control sample 1.}

\item{txBamFile2}{Path to the BAM file for treatment sample 2.}

\item{inBamFile2}{Path to the BAM file for control sample 2.}

\item{span}{The span used for generating Lg2FC values.}
}
\value{
A ComplexHeatmap object representing the log2 fold change heatmap.
}
\description{
Generates a Complex Heatmap of log2 fold change (Lg2FC) values for genomic domains.
}
\examples{
heatmap <- makeDomainHeatPlot(domainsBed = "path/to/domains.bed",
                              windowSize = 100000,
                              stepSize = 5000,
                              txBamFile1 = "path/to/txBam1.bam",
                              inBamFile1 = "path/to/inBam1.bam",
                              txBamFile2 = "path/to/txBam2.bam",
                              inBamFile2 = "path/to/inBam2.bam",
                              span = 3000000)

}
