% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GVizFunctions.R
\name{makeLog2FcCoverageDataTrack}
\alias{makeLog2FcCoverageDataTrack}
\title{Create a DataTrack for Log2 Fold Change (Log2FC) Coverage Data}
\usage{
makeLog2FcCoverageDataTrack(
  chomosome = "chr3",
  start = 5e+06,
  end = 5.5e+07,
  windowSize = 25000,
  stepSize = 5000,
  txBamFile =
    "../01_Asynchronous_HCT116/results/mergedDownSampledBams/TICRR_WT_HCT116_anti-GFP_250.bam",
  inBamFile =
    "../01_Asynchronous_HCT116/results/mergedDownSampledBams/WT_HCT116_anti-GFP_250.bam",
  trackName = "Log2FC",
  HistogramColor = "#1b9e77"
)
}
\arguments{
\item{chomosome}{The chromosome for the genomic region of interest.}

\item{start}{The start position of the genomic region.}

\item{end}{The end position of the genomic region.}

\item{windowSize}{The size of the sliding windows for Log2FC calculation.}

\item{stepSize}{The step size for sliding windows.}

\item{txBamFile}{The path to the treatment BAM file.}

\item{inBamFile}{The path to the control BAM file.}

\item{trackName}{The name for the DataTrack.}

\item{HistogramColor}{The color for the histogram in the DataTrack.}
}
\value{
A DataTrack object for Log2FC coverage data.
}
\description{
This function generates a DataTrack for Log2FC coverage data based on
input BAM files. It calculates Log2FC values for specified genomic regions
and creates a DataTrack for visualization.
}
\examples{
\dontrun{
# Example usage:
dt <- makeLog2FcCoverageDataTrack(chomosome="chr3", start=5000000, end=55000000,
                                  windowSize=25000, stepSize=5000,
                                  txBamFile="path/to/txBamFile.bam",
                                  inBamFile="path/to/inBamFile.bam",
                                  trackName="Log2FC", HistogramColor="#1b9e77")
}

}
