% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeLog2FcGenomePlot.R
\name{makeLog2FcGenomePlot}
\alias{makeLog2FcGenomePlot}
\title{Generate a Genome-Wide Log2 Fold Change Plot}
\usage{
makeLog2FcGenomePlot(
  chrom = "chr3",
  trackStart = 5e+06,
  trackEnd = 5.5e+07,
  windowSize = 25000,
  stepSize = 5000,
  txBamFile =
    "../01_Asynchronous_HCT116/results/mergedDownSampledBams/TICRR_WT_HCT116_anti-GFP_250.bam",
  inBamFile =
    "../01_Asynchronous_HCT116/results/mergedDownSampledBams/WT_HCT116_anti-GFP_250.bam",
  yLimits = c(-0.8, 0.8),
  autoCalculateYLimits = TRUE,
  fillColor = "#006494",
  tx_scaleFactor = NULL,
  in_scaleFactor = NULL,
  blacklist_granges = NULL
)
}
\arguments{
\item{chrom}{Character. Chromosome to analyze (e.g., `"chr3"`).}

\item{trackStart}{Integer. Genomic start coordinate of the region to analyze.}

\item{trackEnd}{Integer. Genomic end coordinate of the region to analyze.}

\item{windowSize}{Integer. Size of the rolling window for smoothing signal (default: 25000).}

\item{stepSize}{Integer. Step size between sliding windows (default: 5000).}

\item{txBamFile}{Character. Path to the treatment BAM file.}

\item{inBamFile}{Character. Path to the control/input BAM file.}

\item{yLimits}{Numeric vector of length 2, defining the y-axis limits (default: c(-0.8, 0.8)).}

\item{autoCalculateYLimits}{Logical. If TRUE, compute y-axis limits automatically from data (default: TRUE).}

\item{fillColor}{Character. Fill color for the area plot (default: "#006494").}

\item{tx_scaleFactor}{Numeric or NA. Optional library size scaling factor for treatment BAM (default: NULL).}

\item{in_scaleFactor}{Numeric or NA. Optional library size scaling factor for input BAM (default: NULL).}

\item{blacklist_granges}{GenomicRanges object with blacklisted regions (default:  NULL)}
}
\value{
A `ggplot2` object showing the log2 fold change across the specified region.
}
\description{
This function computes the log2 fold change (log2FC) in read coverage between treatment and input BAM files
over a specified genomic region, and visualizes it as an area plot using a sliding window approach.
}
\examples{
\dontrun{
makeLog2FcGenomePlot(
  chrom = "chr3",
  trackStart = 1e6,
  trackEnd = 2e7,
  yLimits = c(-1, 1),
  fillColor = "steelblue"
)
}
}
