% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_functions.R
\name{makeLog2FcGenomePlot_2SampleMean}
\alias{makeLog2FcGenomePlot_2SampleMean}
\title{Plot Log2 Fold Change Between Two Sample Pairs Using Mean CPM}
\usage{
makeLog2FcGenomePlot_2SampleMean(
  chrom = "chr3",
  trackStart = 5e+06,
  trackEnd = 5.5e+07,
  windowSize = 25000,
  stepSize = 5000,
  txBamFile1 =
    "../01_Asynchronous_HCT116/results/mergedDownSampledBams/TICRR_WT_HCT116_anti-GFP_250.bam",
  inBamFile1 =
    "../01_Asynchronous_HCT116/results/mergedDownSampledBams/WT_HCT116_anti-GFP_250.bam",
  txBamFile2 =
    "../01_Asynchronous_HCT116/results/mergedDownSampledBams/TICRR_WT_HCT116_anti-GFP_1000.bam",
  inBamFile2 =
    "../01_Asynchronous_HCT116/results/mergedDownSampledBams/WT_HCT116_anti-GFP_1000.bam",
  yLimits = c(-0.8, 0.8),
  fillColor = "#006494"
)
}
\arguments{
\item{chrom}{Character. Chromosome to analyze (default: `"chr3"`).}

\item{trackStart}{Integer. Start coordinate of the genomic region (default: 5000000).}

\item{trackEnd}{Integer. End coordinate of the genomic region (default: 55000000).}

\item{windowSize}{Integer. Size of the window used for rolling sum (default: 25000).}

\item{stepSize}{Integer. Step size between sliding windows (default: 5000).}

\item{txBamFile1}{Character. Path to the first treatment BAM file.}

\item{inBamFile1}{Character. Path to the first input/control BAM file.}

\item{txBamFile2}{Character. Path to the second treatment BAM file.}

\item{inBamFile2}{Character. Path to the second input/control BAM file.}

\item{yLimits}{Numeric vector. Limits for the y-axis (default: c(-0.8, 0.8)).}

\item{fillColor}{Character. Fill color for the area plot (default: "#006494").}
}
\value{
A `ggplot2` object visualizing the log2 fold change between average treatment and control CPMs
        across the genomic region.
}
\description{
This function computes the average CPM (counts per million) across two treatment and two control BAM files
within a specified genomic region, calculates the log2 fold change (log2FC), and plots the result.
The signal is smoothed using a rolling window and visualized as a filled area plot.
}
\examples{
\dontrun{
makeLog2FcGenomePlot_2SampleMean(
  chrom = "chr3",
  trackStart = 1e6,
  trackEnd = 2e7,
  txBamFile1 = "treatment_250.bam",
  inBamFile1 = "input_250.bam",
  txBamFile2 = "treatment_1000.bam",
  inBamFile2 = "input_1000.bam"
)
}
}
