% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeOverlapThresholdSummary.R
\name{makeOverlapThresholdSummary}
\alias{makeOverlapThresholdSummary}
\title{Generate a data frame of overlap counts across replicate peak sets}
\usage{
makeOverlapThresholdSummary(
  reproduciblePeaksPath = "../MakeReproduciblePeaks/MTBP_Peaks_Reproducible_6.narrowPeak",
  replicatePeakPaths = paste0("../01_Asynchronous_HCT116/results/macs2_normalPeaks/",
    asynchronousPeakFiles[grep("MTBP", asynchronousPeakFiles)]),
  targetName = "MTBP"
)
}
\arguments{
\item{reproduciblePeaksPath}{Character string. File path to a BED file containing reproducible peaks.}

\item{replicatePeakPaths}{Character vector. File paths to BED files of replicate peak calls.}

\item{targetName}{Character string. Label to associate with the resulting data (e.g., \"MTBP\", \"TRESLIN\").}
}
\value{
A data frame with columns: \code{peakCount}, \code{peakOverlapThreshold}, \code{Target}, and \code{peakPercentage}.
}
\description{
This function calculates how many peaks from a reproducible peak set are supported by at least
a given number of replicate peak sets. It returns a data frame where each row corresponds to
a different overlap threshold.
}
\examples{
makeOverlapThresholdSummary(
  reproduciblePeaksPath = "MTBP_Peaks_Reproducible_6.narrowPeak",
  replicatePeakPaths = c("rep1.bed", "rep2.bed", "rep3.bed"),
  targetName = "MTBP"
)

}
