% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_functions.R
\name{makePlotFromBeds}
\alias{makePlotFromBeds}
\title{Create a plot from BED files}
\usage{
makePlotFromBeds(
  BedFilenames,
  plotChoice,
  sampleLabelsDF,
  colrs,
  alfa,
  fontScale,
  minFracOverlap
)
}
\arguments{
\item{BedFilenames}{A character vector containing paths to the BED files.}

\item{plotChoice}{A character string specifying the type of plot to create ("UpSet" or "Euler").}

\item{sampleLabelsDF}{A data frame with file labels and corresponding file paths.}

\item{colrs}{A list of colors for the plot elements.}

\item{alfa}{Alpha transparency value for plot elements.}

\item{fontScale}{Scaling factor for text size in the plot.}

\item{minFracOverlap}{Minimum fraction of overlap required for inclusion in the plot.}
}
\value{
NULL (The function generates and displays the plot without returning a value.)
}
\description{
This function reads a list of BED files, calculates overlaps, and creates a plot
based on the specified plot choice (UpSet or Euler). It can be used to visualize
genomic overlap patterns between the BED files.
}
\examples{
BedFilenames <- c("path/to/bedfile1.bed", "path/to/bedfile2.bed")
plotChoice <- "UpSet"
sampleLabelsDF <- data.frame(files = BedFilenames, labels = c("Label1", "Label2"))
colrs <- list(colors = c("#1b9e77", "#d95f02"))
alfa <- 0.7
fontScale <- 8L
minFracOverlap <- 0L
makePlotFromBeds(BedFilenames, plotChoice, sampleLabelsDF, colrs, alfa, fontScale, minFracOverlap)

}
