% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parseUCSCtoGRanges.R
\name{parseUCSCtoGRanges}
\alias{parseUCSCtoGRanges}
\title{Convert UCSC-style coordinate strings to a GRanges object}
\usage{
parseUCSCtoGRanges(coord_strings, genome = "unknown")
}
\arguments{
\item{coord_strings}{A character vector of UCSC-style coordinates, each in the format "chr:start-end".}

\item{genome}{A character string indicating the genome build (e.g., "hg38", "mm10"). Default is "unknown".}
}
\value{
A \code{GRanges} object with one range per input string.
}
\description{
Parses UCSC-style genome coordinate strings (e.g., "chr1:10000-20000")
and returns a corresponding \code{GRanges} object.
}
\examples{
ucsc_coords <- c("chr1:10000-20000", "chr2:15000-15500")
gr <- parseUCSCtoGRanges(ucsc_coords, genome = "hg38")
gr

}
