% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_functions.R
\name{plot_peak_annotation_proportions}
\alias{plot_peak_annotation_proportions}
\title{Plot Proportion of Genomic Features by Peak Set}
\usage{
plot_peak_annotation_proportions(annotation_df, fill_colors)
}
\arguments{
\item{annotation_df}{Combined annotation data frame.}

\item{fill_colors}{Named vector of fill colors for genomic features.}
}
\value{
A ggplot object showing stacked bar chart of annotation percentages.
}
\description{
Plot Proportion of Genomic Features by Peak Set
}
