% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_functions.R
\name{readBed}
\alias{readBed}
\title{Read and convert a BED file to a GenomicRanges object}
\usage{
readBed(bedFile)
}
\arguments{
\item{bedFile}{Path to the BED file to be read and converted.}
}
\value{
A GenomicRanges object representing the data in the BED file.
}
\description{
This function reads a BED file and converts it into a GenomicRanges object
using the GenomicRanges package. The BED file should have columns named
"seqnames", "start", and "end" to represent genomic ranges.
}
\examples{
bedFile <- "path/to/your/bedfile.bed"
gr <- readBed(bedFile)
}
