% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sansam_ggplot_themes.R
\name{sansam_theme}
\alias{sansam_theme}
\title{Sansam Lab Default ggplot2 Theme}
\usage{
sansam_theme()
}
\value{
A \code{ggplot2::theme} object with customized appearance:
\itemize{
\item Transparent background
\item Small font sizes (6–8 pt)
\item No outer margins
}
}
\description{
A minimalist \code{ggplot2} theme with small font sizes and transparent backgrounds,
designed for compact scientific figures.
}
\examples{
library(ggplot2)
ggplot(mtcars, aes(mpg, wt)) +
  geom_point() +
  sansam_theme()

}
