% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setCustomGvizScheme.R
\name{setCustomGvizScheme}
\alias{setCustomGvizScheme}
\title{Apply Custom Gviz Scheme for Consistent Track Styling}
\usage{
setCustomGvizScheme()
}
\value{
Invisibly returns the name of the applied scheme.
}
\description{
This function defines and registers a customized Gviz visual scheme called `"myScheme"`,
which sets consistent font sizes, colors, and graphical scaling for genome tracks.
It then sets it as the default Gviz scheme via `options(Gviz.scheme = "myScheme")`.
}
