% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treslin_peak_granges.R
\docType{data}
\name{treslin_peak_granges}
\alias{treslin_peak_granges}
\title{GenomicRanges of TRESLIN Cut&Run Peaks}
\format{
A named list of GenomicRanges objects.
}
\source{
GEO: GSE294620, GSE274463, GSE274464, GSE274465
}
\usage{
data(treslin_peak_granges)
}
\description{
A list of GenomicRanges representing TRESLIN peaks used in Figure 1 of Noble et al. (2024) https://doi.org/10.1101/2024.02.02.578516.
}
\details{
All peaks are called with MACS2 and used to derive reproducible peaks in the analysis.
}
\keyword{datasets}
