% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hg38_ideogram_granges.R,
%   R/treslin_peak_granges.R
\docType{data}
\name{hg38_ideogram_granges}
\alias{hg38_ideogram_granges}
\title{GenomicRanges of hg38 ideogram}
\format{
A GenomicRanges object.

A GenomicRanges object.
}
\usage{
data(hg38_ideogram_granges)

data(hg38_ideogram_granges)
}
\description{
GenomicRanges of hg38 ideogram

GenomicRanges of hg38 ideogram
}
\keyword{datasets}
