% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_df_of_genomic_features_overlapping_peaks.R
\name{make_df_of_genomic_features_overlapping_peaks}
\alias{make_df_of_genomic_features_overlapping_peaks}
\title{Annotate Genomic Features Overlapping Peaks}
\usage{
make_df_of_genomic_features_overlapping_peaks(
  peaks_gRanges,
  annotationDb = "org.Hs.eg.db",
  TxDb = txdb,
  SetName = "All Peaks"
)
}
\arguments{
\item{peaks_gRanges}{A \code{GRanges} object containing genomic peak locations.}

\item{annotationDb}{Character. The annotation database for gene annotations. Defaults to \code{"org.Hs.eg.db"} for human genes.}

\item{TxDb}{A \code{TxDb} object specifying transcript annotations. Defaults to \code{txdb}.}

\item{SetName}{Character. The name of the peak set (used for labeling in downstream analyses). Defaults to \code{"All Peaks"}.}
}
\value{
A data frame summarizing the genomic features overlapping the given peaks.
The data frame includes:
\itemize{
\item \code{Feature}: Genomic feature (e.g., "Promoter", "Exon", "Intron").
\item \code{Frequency}: The number of peaks overlapping each feature.
\item \code{Set}: The label for the peak set.
}
}
\description{
This function annotates genomic regions overlapping with peaks using ChIPseeker.
}
\examples{
library(GenomicRanges)
peaks <- GRanges("chr1", IRanges(start = c(100100, 150000, 180000), width = 200))
txdb <- TxDb.Hsapiens.UCSC.hg19.knownGene
df <- make_df_of_genomic_features_overlapping_peaks(peaks, TxDb = txdb)
head(df)
}
\seealso{
\code{\link{make_stacked_barchart_of_genomic_features_overlapping_peaks}}
}
